rename duplicated short name to DonwgradeTypedPropertyRector
