+++ config/set/downgrade.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use Rector\Downgrade\Rector\Property\TypedPropertyRector;
+use Rector\Downgrade\Rector\Property\DowngradeTypedPropertyRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
 
-    $services->set(TypedPropertyRector::class);
+    $services->set(DowngradeTypedPropertyRector::class);
 };
