+++ rules/downgrade/tests/Rector/Property/DowngradeTypedPropertyRector/Fixture/classname_nullable_type.php.inc
@@ -1,0 +1,22 @@
+<?php
+
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
+
+class ClassNameNullableTypeClass {
+    private ?\Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Source\AnotherClass $property;
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
+
+class ClassNameNullableTypeClass {
+    /**
+     * @var \Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Source\AnotherClass|null
+     */
+    private $property;
+}
+
+?>
