+++ rules/downgrade/tests/Rector/Property/DowngradeTypedPropertyRector/Fixture/fixture.php.inc
@@ -1,0 +1,22 @@
+<?php
+
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
+
+class SomeClass {
+    private string $property;
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
+
+class SomeClass {
+    /**
+     * @var string
+     */
+    private $property;
+}
+
+?>
