+++ rules/downgrade/tests/Rector/Property/DowngradeTypedPropertyRector/Fixture/nullable_type.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
 
 class NullableTypeClass {
     private ?string $property;
@@ -10,7 +10,7 @@ class NullableTypeClass {
 -----
 <?php
 
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Fixture;
 
 class NullableTypeClass {
     /**
