+++ rules/downgrade/tests/Rector/Property/DowngradeTypedPropertyRector/Source/AnotherClass.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Rector\Downgrade\Tests\Rector\Property\DowngradeTypedPropertyRector\Source;
+
+
+class AnotherClass
+{
+
+}
