--- rules/downgrade/tests/Rector/Property/TypedPropertyRector/Fixture/classname_nullable_type.php.inc
@@ -1,22 +1,0 @@
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class ClassNameNullableTypeClass {
-    private ?\Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Source\AnotherClass $property;
-}
-
-?>
------
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class ClassNameNullableTypeClass {
-    /**
-     * @var \Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Source\AnotherClass|null
-     */
-    private $property;
-}
-
-?>
