--- rules/downgrade/tests/Rector/Property/TypedPropertyRector/Fixture/classname_type.php.inc
@@ -1,22 +1,0 @@
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class ClassNameClass {
-    private \Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Source\AnotherClass $property;
-}
-
-?>
------
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class ClassNameClass {
-    /**
-     * @var \Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Source\AnotherClass
-     */
-    private $property;
-}
-
-?>
