--- rules/downgrade/tests/Rector/Property/TypedPropertyRector/Fixture/fixture.php.inc
@@ -1,22 +1,0 @@
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class SomeClass {
-    private string $property;
-}
-
-?>
------
-<?php
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Fixture;
-
-class SomeClass {
-    /**
-     * @var string
-     */
-    private $property;
-}
-
-?>
