--- rules/downgrade/tests/Rector/Property/TypedPropertyRector/Source/AnotherClass.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace Rector\Downgrade\Tests\Rector\Property\TypedPropertyRector\Source;
-
-
-class AnotherClass
-{
-
-}
